assignment-3/requirements.txt
about-time==4.2.1
alive-progress==3.2.0
asttokens==3.0.0
autograd==1.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
cma==3.2.2
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
deprecated==1.2.18
dill==0.4.0
executing==2.2.0
fonttools==4.57.0
grapheme==0.6.0
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
joblib==1.4.2
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.1
matplotlib-inline==0.1.7
mealpy==3.0.1
narwhals==1.38.2
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
opfunu==1.0.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
platypus-opt==1.4.1
plotly==6.0.1
prompt-toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
pymoo==0.6.1.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
wrapt==1.17.2
